+++ src/Components/Components/test/RenderTreeDiffBuilderTest.cs
@@ -30,9 +30,9 @@ namespace Microsoft.AspNetCore.Components.Test
         void IDisposable.Dispose()
         {
             renderer.Dispose();
-            oldTree.Dispose();
-            newTree.Dispose();
-            batchBuilder.Dispose();
+            ((IDisposable)oldTree).Dispose();
+            ((IDisposable)newTree).Dispose();
+            batchBuilder?.Dispose();
         }
 
         [Theory]
