+++ src/Components/Components/test/Rendering/ArrayBuilderSegmentTest.cs
@@ -32,7 +32,7 @@ namespace Microsoft.AspNetCore.Components.Rendering
         public void StillWorksAfterUnderlyingCapacityChange()
         {
             // Arrange: builder containing 1..8
-            using var builder = new ArrayBuilder<int>(minCapacity: 10);
+            using var builder = new ArrayBuilder<int>(minCapacity: 10, new TestArrayPool<int>());
             builder.Append(new[] { 1, 2, 3, 4, 5, 6, 7, 8 }, 0, 8);
             var originalBuffer = builder.Buffer;
